Add configuration files and scripts for WSJ and chiangyihan egs; implement various utility scripts

+++ egs/chiangyihan/s5/run_combine_augment_JnrleV1.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 set -x
-set +e
+set -e
+export PS4='+[${BASH_SOURCE}:${LINENO}] '
+
 . ./path.sh
 . ./cmd.sh
 # Copyright 
@@ -25,16 +27,17 @@ Step10_TDNN_Pre=1
 Step11_TDNN_train=1
 Step12_Decoding=1
 
-model_name=Wang_real_addnoise_LDV_caffeteria_n55_LDV_s40
+
+
 
 #路徑cd到S5
 mfcc_conf_path=/work/jerryfat/kaldi-trunk/egs/chiangyihan/s5/data/LDV_VITS_asr/conf_forfun/mfcc_hires.conf
 s5_root=$(pwd)
+model_name=Wang_real_addnoise_LDV_caffeteria_n55_LDV_s40
 data=$(pwd)/data/LDV_asr_Jnrle_v1/TEST/${model_name}/final #change
 data_root=$(pwd)/data/LDV_asr_Jnrle_v1/TEST/${model_name} 
 nj=2 #CPU 線程
 passward=qwertyuiop #使用者密碼
-
 
 #獲取當前路徑存到data
 echo "當前路徑為$data"
@@ -48,24 +51,38 @@ echo "當前路徑為$data"
 # utils/fix_data_dir.sh $data_root
 # echo -e "#####End data preparation#####"
 #data preparation end ####################################
+export USE_SEGMENTS=false  # 全域設定,讓所有腳本都知道不使用 segments
+if [ -f $data_root/raw/segments ]; then
+    rm $data_root/raw/segments
+fi
 
 if [ $Step1_Data_Aug -eq 1 ]; then
 echo -e "#####Start Augmentatiom preparation#####"
 cd /work/jerryfat/kaldi-trunk/egs/chiangyihan/s5
+
+# 先刪除可能存在的 segments 文件
+rm -f $data_root/raw/segments 2>/dev/null || true
+
+# 確保 wav.scp 存在且有內容
+if [ ! -f $data_root/raw/wav.scp ]; then
+    echo "Error: wav.scp not found in $data_root/raw"
+    exit 1
+fi
+
+# 計算原始時長（不使用 segments 文件）
+utils/data/get_utt2dur.sh --nj 2 $data_root/raw
+
+# 繼續原有的處理
 utils/data/perturb_data_dir_speed_3way.sh $data_root/raw $data_root/speed
 steps/nnet2/get_perturbed_feats.sh --feature-type mfcc conf/mfcc_hires.conf $data_root/perturb/mfcc $data_root/perturb/mfcc_perturb $data_root/speed $data_root/perturb/mfcc_dataset
 # steps/nnet2/get_perturbed_feats.sh --feature-type mfcc $mfcc_conf_path $data_root/perturb/mfcc $data_root/perturb/mfcc_perturb $data_root/speed $data_root/perturb/mfcc_dataset
 
-echo -e "#####removeing_frame_shift#####"
 rm $data_root/perturb/mfcc_dataset-0.9-0.9/frame_shift
 rm $data_root/perturb/mfcc_dataset-0.95-1.1/frame_shift
 rm $data_root/perturb/mfcc_dataset-1.0-0.8/frame_shift
 rm $data_root/perturb/mfcc_dataset-1.05-1.2/frame_shift
 rm $data_root/perturb/mfcc_dataset-1.1-1.0/frame_shift
-echo -e "#####removeing_frame_shift_end#####"
-
-
-
+	
 #combine mfcc augmentaion
 utils/data/combine_data.sh $data_root/perturb/mfcc_dataset $data_root/perturb/mfcc_dataset-0.9-0.9 $data_root/perturb/mfcc_dataset-0.95-1.1 $data_root/perturb/mfcc_dataset-1.0-0.8 $data_root/perturb/mfcc_dataset-1.05-1.2 $data_root/perturb/mfcc_dataset-1.1-1.0
 
@@ -96,17 +113,15 @@ fi
 if [ $Step3_Lexicon_Pre -eq 1 ]; then
 ### dict Start###
 echo -e "#####Start lexicon preparation#####"
-echo "當前路徑為$(pwd)"
 cd ../ #cd local
 mkdir -p dict_learning
 cp $data_root/lexicon.txt dict_learning/lexicon.txt || exit 1
-dos2unix dict_learning/lexicon.txt || exit 1 #change
+dos2unix /work/jerryfat/kaldi-trunk/egs/chiangyihan/s5/data/cva_asr/data/test/local/dict_learning/lexicon.txt || exit 1 #change
 cd dict_learning
 echo SIL > silence_phones.txt || exit 1
 echo SIL > optional_silence.txt || exit 1
 grep -v -w sil lexicon.txt | awk '{for(n=2;n<=NF;n++) { p[$n]=1; }} END{for(x in p) {print x}}' | sort > nonsilence_phones.txt || exit 1 
-echo "當前路徑為$(pwd)"
-dos2unix nonsilence_phones.txt || exit 1  #change
+dos2unix /work/jerryfat/kaldi-trunk/egs/chiangyihan/s5/data/cva_asr/data/test/local/dict_learning/nonsilence_phones.txt || exit 1  #change
 echo -e "<unk>\tSIL" >> lexicon.txt  || exit 1
 cd $s5_root #cd S5
 
@@ -123,14 +138,17 @@ if [ $Step4_Subset_Aug_Pre -eq 1 ]; then
 # utils/subset_data_dir.sh $data_root/raw 1000 $data/inside_test || exit 1 
 utils/subset_data_dir.sh $data_root/raw 50 $data/inside_test || exit 1 
 
-dos2unix $data/text || exit 1               #change
-dos2unix $data/inside_test/text || exit 1    #change
+# dos2unix $data/text || exit 1               #change
+# dos2unix $data/inside_test/text || exit 1    #change
 
 utils/fix_data_dir.sh $data
 # 製作Over
 fi
 
 if [ $Step5_MFCC_Extract -eq 1 ]; then
+
+# 確保有 utt2dur
+utils/data/get_utt2dur.sh --nj $nj --segments false $data
 
 ### features extraction###
 echo -e "#####Compute MFCC for training#####"
@@ -220,7 +238,7 @@ fi
 
 if [ $Step9_Subset_final_Pre -eq 1 ]; then
 ##製作Subset for inside test
-utils/subset_data_dir.sh $data/nopitch 150 $data/nopitch_inside || exit 1
+utils/subset_data_dir.sh $data/nopitch 250 $data/nopitch_inside || exit 1
 steps/online/nnet2/extract_ivectors_online.sh --cmd run.pl --nj $nj $data/nopitch_inside exp_learning/$model_name/nnet3/extractor $data/nopitch_inside/ivectors || exit 1
 fi
 
@@ -369,4 +387,3 @@ echo -e "#####Inside testing#####\n"
 # echo -e "#####Outside testing#####\n"
 
 fi
-
